requirements.txt
ccxt>=4.5.4
eth-abi>=5.2.0
eth-utils>=5.3.1
flask>=3.1.2
matplotlib>=3.10.6
python-telegram-bot>=22.3
requests>=2.32.5
web3>=7.13.0
base58
solana
TA-Lib==0.6.7
numpy==2.3.3
pandas==2.3.2
fastapi
uvicorn
python-dotenv
gunicorn==21.2.0
aiohttp
telegram
